Make conversations repository to emit nullable values when conversation is not found.
It has been changed as this seems to be the way the store library works, i.e. when it received a null from the db it can try to fetch the data from the remote. If in received an exception from the DB/room it will just rethrow this exception to the output flow (which we do not want).

MAILAND-1968