Fix issue where "EMPTY_MESSAGE" would be used instead of quoted body when replying
The root cause of this issue was that once the message body was
decrypted and formatted, it was not posted back to the
`decryptedConversationUiModel` which is later queried for the decrypted
message body HTML when the reply action is triggered.

As a solution, we post it back on the live data for both the "normal"
case of decrypting a message (through `formatMessageHtmlBody`) and the
one where the user requires rendering of embedded images (thorugh
observing of `renderedMessage`) as that operation also involves doing
changes to the decrypted message HTML body which are then later needed
when replying.

Please note that the "render embedded images" case was already
working as we were indeed setting the decryptedHTML on the live data's
message, which was enough to have the "reply" listener accessing this
value through reference: this is very confusing and shouldn't be relied
upon, so as a minumum we want to have a "post" call to this live data in
place.

MAILAND-2009