Remove jacoco dependency to avoid filling the unit tests logs with exceptions
Unit Tests runs started throwing exceptions as the following
```
ava.lang.IllegalStateException: Cannot process instrumented class
<class>. Please supply original non-instrumented classes
```
many times for each test, causing the logs to get filled and unreadable.

While there are several pointers to the potential causes online we
didn't find a solution to get rid of them in a short timebox, so since
`jacoco` is currently not being used we decided to disable it for now.