Fixed conversation details not appearing after selecting.
There was a problem that when a conversation with multiple messages is loaded and used click on one of closed messages, the message content is not displayed for the first time. This commit fixes it. But it does not fix the underlying problem which is a mutability of Message object.

MAILAND-1968