Fix mark read/unread and star/unstar in conversations.
The actions above were not working from action sheet due to the fact that the type of conversation/message was being resolved using only last message from the conversation (which could have been in the location indicated as forced single type). What was changed is that in case of details conversation action sheet the location should be ignored.

MAILAND-2074