Fix null pointer exception in `AttachmentClearingService.java`
A null pointer exception was triggered when trying to call
`attachmentMetadataDao.getAllAttachmentsSizeUsedBlocking()`
which has a return type `Long?` and trying to cast it to a `long` in java.

We add another function
`attachmentMetadataDao.getAllAttachmentsSizeUsedNonNullBlocking()`
that interprets a null value as zero.

We modify the service to call this one instead.