Do not treat "Message Already Sent" as an error when sending a message
This is to avoid some cases where errors would be displayed to users
when there was no actual issue.
One such case might happen when killing the app right after sending a
message:
If the "send message" API request was already fired but the worker
didn't complete yet, when the worker will be restarted by the OS it
will enter a failure loop as the save draft use case will fail with
"Message Already sent" error.

At this point, we handle the error by propagating it up to the send
worker, which will ensure the given message is moved out of
drafts to the send folder and then will log this error and finish.

MAILAND-2010