Fix problems when moving messages to Trash folder
There was a crash when moving a starred message to trash. The reason was that after removing the current location from the list of labels, we were trying to resolve the location based on the remaining label ids (5 and 10), from which a location cannot be resolved. The solution was to add the destination location to the list of labels before the current location is removed. Also, when removing "old folder ids" we need to exclude the destination location from the removal i.e. trash.

MAILAND-2124