Solve 'load remote content' inconsistencies
This addresses two issues:
- 'Load remote content' button sometimes not shown when needed
- Remote content sometimes not loaded when 'auto load remote content'
was enabled and 'load remote content' button shown instead

The first issue was caused by `loadDataWithBaseURL` being called more
than once, sometimes before the data was ready: this caused webview to
misbehave and `onPageFinished` callback to not being called for all
pages loaded - which consequentially caused the button not to be shown.
The solution was to remove the uneeded calls to `loadDataWithBaseURL`.

The second issue was caused by the `PMWebViewClient` being initialised
with "false" as 'loadRemote' parameter.
This caused the webview to always block the remote content at the first load,
which as a direct consequence caused the "load remote content" button to
show and the content to not being shown automatically as the webview was
not reloaded.

MAILAND-2049