Solve issue with message sometimes being blank when loading remote content
When having the 'auto load remote content' feature disabled and
pressing 'load remote content' button for a message, the message would
sometimes "disappear". This was due to the webview size being changed to 0dp
during the "reload" operation, for causes not completely understood yet.
Forcing the webview to be redrawn after reloading its content have shown
to prevent this in the tested cases.

MAILAND-2049