Catch and handle exception when parsing label colors
An exception is thrown sometimes when opening the labels action sheet and the app crashes. The reason is an invalid color value passed to the toColorInt() function. Because we cannot reproduce and pinpoint the core problem, for now we catch this exception, log it, and set Color.GRAY as a fallback when we cannot parse the color.

MAILAND-2135