Fix not being able to create and update labels with 2 of the colors
There's a list of allowed colors for the labels on the backend. Two of the colors that were in the designs are not present in this list. Because of this we had done a workaround to map the colors shown in our app to the colors that are actually allowed on the backend. Apart from the workaround, this situation still caused problems in the app sometimes, like creating and updating labels. In this commit the workaround is removed and the two problematic colors are replaced with the colors that are actually allowed by the backend, which are very similar to the ones in the designs.

MAILAND-2156