Solve issue when forwarding an old message from search list
Sending was failing when trying to forward a message from search list.
This was happening when the message was old enought to not be available
in the messagesDatabase anymore, and was caused by the fact that sending
process doesn't take into account the "searchDatabase" when searching
for the parent message (as such DB is scheduled for removal).

The solution is to fetch the parent message from the API when not
available locally.

- Removing the `parentMessage?.attachments(database)` method call as not
  needed, since the same is performed by
  `messageDetailsRepository.findMessageByIdBlocking()`. This was a
  leftover from the old logic which was performing this additional call
  as it was choosing which DB to pass in dynamically (messages / search)

MAILAND-2010