Refresh displayed composer attachments when opening AddAttachmentsActivity
after a draft was created.

This is needed to solve the issue detailed in MAILAND-2010/comment-84445:
When forwarding a message from search, if the attachments screen was
opened after the draft creation happened it would contain outdated
attachmentIds, causing the removal of attachments to fail.

This fix ensures that the displayed attachments represent what is
available in the DB in this moment.

MAILAND-2010