Fix mark unread action in conversation mode not working in custom folders
The API requires for us to send a location id from where we're performing the action. When we try to send one of the locations that are part of MessageLocationType we can end up sending a location that is not actually valid for the backend (example: LABEL, LABEL_FOLDER). That's why we need to send an additional parameter to the ConversationsRepository that will be a string and will represent the location id. That way we can send the id of a custom location when we perform the action from some custom location. So, when we're in a custom location, this id will be the id of that location, otherwise it will be a string representation of the location from MessageLocationType.

MAILAND-2074