Fix crash when moving starred messages from any folder to inbox, archive or custom folder
The problem is that we first remove the current location and then we add the new one. On each addLabels or removeLabels call the resolver tries to resolve the location of the message, so in this scenario the resolver tries to resolve the location when there is no valid location in the list. The solution is to reverse the operations. We first need to set the new location and then remove the old one so the crash won't happen.

MAILAND-2170