Solve glitch when collapsing last message after scrolling
This issue was due to the webview size being changed dynamically at bind
time, which caused the adapter to get confused with the positioning of
the header when removing the body for a message (which is a collapsing op).

Notifying the adapter that the header item changed when expanding /
collapsing any message, "hacks" the adapter into re-calculating the
header's position, solving the glitch.

- Move "expand header chevron" and "last message divider" view's
  visibility handling out of the header's click listener to keep it
  consistent over views reciclying.

MAILAND-2084