Solve issue where reply, reply all and forward would apply to the wrong message
This issue happened when performing one of the above actions to a
message in a converation through the "In message actions" action sheet.
The cause is that when executed from the actionsheet, such actions were
always being applied to the last message.

The patch is to pass the messageId on which we want the action to be
applied back to the method that executes the actions on the activity.

MAILAND-2185