Handle the scenario when a conversation with a given ID is not found in the DB when calling read/unread actions
A crash was happening sometimes when calling read/unread actions where the conversation object was required (requireNotNull). With this commit instead of using requireNotNull, we check whether the conversation is null or not and depending on that we return a Success or Error result. The Error result results in showing a toast message that says "Could not complete the action". This error handling is done because the core reason why sometimes the conversation is not found in the DB cannot be identified.

MAILAND-2171