Replying to a conversation replies to the last message which is not a draft
both when executed from "more actions" menu and from "bottom bar".
The same applies for reply all and forward actions.

- Conversation Details expands the last message that is not a draft

Moved the responsibility to decide which message is expanded when
the adapter is inited from the `ExpandableRecyclerAdapter` to
`MessageDetailsAdapter`, to allow checking in the message state and
avoid expanding draft messages.

The above, on top of being the desired behavior, was needed to ensure
that when replying to a conversation the body and `decryptedHtml` fields
of the last message that is not a draft had actually been loaded (as
they are needed to pass the "quoted text" over to the composer for
replying / forwarding).

MAILAND-2199