Make showing the "move to folder" actions depend on ActionSheetTarget
Showing the "move to folder" action should depend on the location only when we're opening the action sheets from the mailbox and when we're opening them for single messages in conversation details. In the case where the action sheets are summoned for the whole conversation in the conversation details, showing the "move to folder" actions shouldn't depend on the location and all actions should be shown except the "move to trash" action which wouldn't be shown if we're opening the conversation from the trash folder.

MAILAND-2074