Add `isDraft` flag to MailboxUiModel
This flag indicates whether a message is a draft when in message mode
or whether a conversation contains only one message which is a draft
when in conversation mode.

The initial purpose of this was to differentiate between draft / non
draft items when they clicked, to decide whether composer or detail
should be opened.

MAILAND-2224