Change ConversationModeEnabled use case so that it can check if conversation mode is enabled for users other than the primary
When we want to check if conversation mode is enabled for users other than the primary, we can provide the use case with the user id of the user for which we want to check this. This parameter is optional and if it's not provided, by default, the current primary user id will be used.

MAILAND-2221