Display draft icon instead of sender initials for draft messages in all mail
It's considered to be a draft both a message that is in either DRAFT or
ALL_DRAFT folder and a conversation which only contains a message that
is a draft. This logic is exposed through the `isDraft` flag of MailboxUiItem

Move the responsibility to show the draft icon in SenderInitialView and
reuse the same for MessageDetailsHeaderView as well.

MAILAND-2224