Revert "Remove `DUMMY_URL_PREFIX` as "http://androidlinksfix.protonmail.com" from MessageDetailsAdapter.kt"
This reverts commit 88de93ca18aa8d9456859249d19b5347d458209a.

Commit 88de93 was an attempt to fix MAILAND-1801, but dropping the
"DUMMY_URL_PREFIX" had some unintended side effects.

One is that the `webview.reload` call performed to load a message's
remote content stops working properly, reloading to a blank message
instead of reloading the existing HTML content.

This bad behavior reloading only happens when the history of a
message was being displayed (MAILAND-2220).

That seem to happen because the default url for `loadDataWithBaseURL` is
"about:blank"

MAILAND-2220