Fix empty list of folders and labels sometimes in the drawer
The problem was appearing when removing an account from the account list. When an account is removed, so is all the data for that account. However, in the NavigationActivity we continued to observe the labels data for the account that has been removed. So, when the labels data was being removed, the observer was triggered and the labels were removed from the drawer. The solution was to remove the observers before staring to observe the labels for the current primary account.

MAILAND-2139