Fixes background dim disappearing for fully expanded action sheet on large screen devices
On the large screen devices, the fully expanded bottom sheet might not take the whole screen.
In that case, the background dim would be removed with the message view still visible in the background,
resulting in the action sheet and the background blending together.
Fixes that issue by checking the action sheet's top position before deciding if dim should be removed.

MAILAND-2205