Scroll to the last message in the list of messages in conversation details
When there was a conversation with a bigger number of messages, when opening that conversation the expanded message was not visible because it was at the very bottom of the list. So, now when we open the conversation we scroll it to the bottom so that the expanded message can be visible. We do this scrolling on the second call of the observer because in most cases that's how many times the observer is called when opening a conversation. It is expected that in some cases this solution won't work because in some cases the observer will be called less/more times.

MAILAND-2290