Reduce time between "send message" operation retries
When any step of the "Send message" operation fails, an exponential
backoff policy was applied. In the worst case scenario (each of the
possible steps of the sending process failing each possibile
attempt) this "retry policy" would imply long delays and cause the
send operation to become too long (and the user to lose the context of
the ongoing operation).

This commit reduces the number of retries for the "save draft" and
"upload attachments" operations (both part of the sending process) to 1,
the one for "send message" to 2 and the backoff delay to 10 seconds for
all ops.

MAILAND-2293