Enable tests checking "delete item" action visibility
The two tests touched by this commit weren't working because of
`location` property being set at VM's init time, causing it to always
have the "default" value declared by the mock instead of the new one
that each test is declaring.

Making the property not lazy works around this issue, as it will be
evaluated each time it's required. The eval is not expensive and in
production the value returned by "saveStateHandle" won't change after
init time, making this workaround safe.

MAILAND-1998
MAILAND-2161