Disables the customise swipe settings snackbar and undo action from the swipe actions in the mailbox
Disables the custimise settings feature as a product decision.

Removes the undo button from the swipe actions, untill the undo actions are actually implemented.

In addition, fixes a related bug
wherein the snackbars after the swipe actions would not be visible at all, after
the mailbox activity is first created, until the mailbox actions bottom bar is shown
for the first time. This seems to be related to the mailboxActionsView, which acts as
the snackbar anchor view, being initially GONE from the layout. Fix it by making it
invisible with 0dp height initially.

MAILAND-2320
MAILAND-2286