Fixed some false positives in ConversationsRepositoryImplTest
As mentioned in `b0c6ecbafa845c`, some tests in this class were
suceeding without being fully executed because of the usage of expected
exceptions (which cause the test to succeed when the exception is
thrown without executing the whole test).

Using turbine to verify both the error and the actual items solved this.

MAILAND-2234