Change the UpdateConversationsLabelsWorker so that the unchecked labels are not sent as data to the worker
The data we are sending to the worker has a limit of 1024 bytes. If that is exceeded an InvalidStateException is thrown saying "Data cannot occupy more than 10240 bytes when serialized". In our case that translates to maximum of 109 labelIds that we can send to the UpdateConversationsLabelsWorker. So I've moved the figuring out which labels are unchecked logic into the worker itself. Message doesn't have this problem because we still use separate jobs for RemoveLabelJob and ApplyLabelJob..

MAILAND-2311