Convert Message.ParsedHeaders model to a kotlin data class
This change was triggered from an investigation that looked into why a
conversation was automatically updated when opened.

The caused turned out to be a failure in the equality check performed by
`distinctUntilChanged` check in DetailVM (conversationUiModel flow).
This failure was caused by the ParsedHeaders equality failing since not
implemented in the existing java class (checking instance).

- MessagesTypesConverter uses kotlin serialization to serialise ParsedHeaders
- Move test checking parsed headers serialization to a unit test
  (instead of android test)
- Delete uneeded ParsedHeadersMatcher
- Bump MessageDB version to 10

MAILAND-2291