Revert "Debounce conversation detail flow to avoid subsequent refreshes"
This reverts commit 5580db55496b0cfa86995a14c9b32be07ed8f467,
removing the 1000 ms debouce when observing conversation flow

This debouce was introduced by `5580db55496b0` as a patch to avoid
updating the UI multiple times when doing batch operations on messages.

It turned out this debouce creates unexpected issues when performing
other operations, such as marking the last message as read when opening a
conversation: with the debouce, the "marked as read" converation is
emitted after a second instead of immediately, causing the conversation
to "refresh" instead of just loading the "read" message directly.

Since it turned out that the change of order in the execution of "unstarring"
operations (previous commit) makes the "refresh" on conversation behave in the
 correct way also without the `debouce`, this commit drops it.

MAILAND-2291