Message labels are updated locally when conversation "starred" changes
This is needed to avoid an inconsistent state where, till the next
"event" is received all the messages of a conversation that has just
been starred or unstarred would have an updated "isStarred" flag but
out-of-dated "labels" (which would still reflect the previous starred
status.

Within other things, this was causing the UI to be refreshed when the
event with the correct labels was retrieved.

MAILAND-2291