Return embedded images with local files if embedded images are already downloaded in HandleEmbeddedImageAttachments
If the embedded images were downloaded, we would return the list of the embedded images but without the localFileName attributes set. This would cause embedded images to not be loaded sometimes even though they have been downloaded. The fix is to return the embedded images list, with the localFileName attribute set for each embedded image if they have been downloaded already.

MAILAND-2324