Provide embeddedImageIds to the startDownloadEmbeddedImagesJob method in MessageDetailsViewModel, instead of taking the list from the message object
The problem was that embeddedImageIds in Message is being set when decryption of the message is done. In the scenario where we click on the button in order to load embedded images, the message object that was passed to the startDownloadEmbeddedImagesJob didn't have the embeddedImageIds field set because when decryption was done it was not done on that particular instance of the Message object. The solution is to save the list of embeddedImageIds in the MessageDetailsListItem for each message at the time of loading/decrypting the message, and then use it when the button for loading embedded images is clicked.

MAILAND-2324