Remove showing error for selecting more labels than it is allowed to be created
The labels threshold for a user should be checked only during label creation and not label selection. So the only time we show an error now on selection is when more than 100 labels are selected. This is changed because of a bug we have that after login an error is shown after 3 selected labels because it considers 3 as the threshold for creating labels, no matter the actual limit.

MAILAND-2307