Makes sure that the expiration time information is preserved when sending a message
The expiration time would get lost along the way when creating a new message:
- first the expiration time would not get copied to the message object from the
_messageDataResult object upon sending,
- then the CreateDraftWorker would overwrite the message in the db with a 0
after receiving a draft from the BE.

The implemented solution fills is in the missing expiration information in the worker
after receiving a response. A better approach, as the next iteration, would be to
make sure we send the expiration time to the BE when creating a draft to start with.

MAILAND-2364