Avoid compulsive refreshes on Mailbox
In past we were not observing Mailbox Items, but fetching a static list, so we had a lots of refreshes put in place in Mailbox. Now this is not needed anymore, but is causing bugs, particularly the whole flow for fetch Mailbox Items is restarted in some cases, letting the ViewModel think that the User, the Location, or the Label id has changed, giving the List the command to scroll at the top.