Fixes the messages not being marked as unred from the detail view
The issue was caused by the following sequence of events:
- message is marked as unred,
- this leads to the details view being closed and mailbox view being shown,
- in the meantime, the message is marked as unread in the local db which leads to new emission,
- there is a time window where the detail view activity is still not destroyed and receives the event
with the message, unread, being emitted, even though the view is no longer visible to the user, and
marking this message as read again.

The solution is to ignore the emitted messages/conversations after the details activity has been paused-
this way the unread message will not be incorrectly marked as read when the user is already seing
the mailbox activity.

MAILAND-2339