Fixes error while opening some of the conversations
The error was related to the labels that some of the messages within a given conversation have-
some of them had only ALL_SENT and ALL_MAIL labels and none could be resolved to a valid location
when creating a message object, which would lead to an exception being thrown and convo not loading.
Fixed by re-adding the ALL_SENT to valid locations, which was initially removed as part of the "send message to
self" fix, as the other fix applied in that MR already took care of that issue (treating sent as label).

MAILAND-2343