Add SENT_TO_EXTERNAL encryption type
This is needed to handle the case of a message sent from a pm user to an
external address (unencrypted).

Such case wasn't working since our API considers all drafts to be "internal"
messages and returns the "Flags" accordingly, while the android app
considers every message which is internal as end2end encrypted, which is
this case doesn't reflect the truth.

This change fixes an issue for which messages sent to external addresses
would wrongly show a "send by you with end to end encryption" tooltip.

MAILAND-1891