The subject line wasn't collapsing on some devices when the text was just single line
So if the text inside the CollapsingToolbarLayout was multiline then collapsing of the subject worked correctly when scrolling. But if it was single like text, then it wasn't working on some devices. This was because the Coordinator Layout was global and for some reason it wasn't picking up on the scroll under this conditions.
I made the global layout Constraint Layout and now the Coordinator only wraps the components needed for the collapse on scroll to work. The bottom bar doesn't need to be part of the Coordinator Layout since it's a static component anyway.

MAILAND-2415