Mitigate NPE in MessageDetailsViewModel.onEmbeddedImagesDownloaded
This should mitigate the NPE generated in the mentioned tickets. This should be happening when embedded images are downloaded when the Activity is already closed, so ignoring the absence of value and terminate the process would not negatively affect the users.

Worth to mention is that this should not be happening in a sane LifeCycle, as the `viewModelScope` should be already cancelled when this is happening

MAILAND-2332
MAILAND-2394