Changed the way we use the PendingIntent tied to the new message notifications
This issue was reproducible only on some device brands like Huawei and OnePlus. So the Pending Intent, the way we had it implemented was supposed to always be the same event but be updated for each new message, because the extra data changed. And this behaviour was supposed to be handled by the `PendingIntent.FLAG_UPDATE_CURRENT` flag. But with the forementioned manufacturers the event was never replaced.
So the way that I found, to actually always make sure that the intent called was gonna be unique, is to get the intent with a unique requestCode each time.
And also now `Intent.FLAG_ACTIVITY_NO_HISTORY` on the MessageDetails intent fixes the original issue, and because it makes sure we don't save the state of the activity we always reload the data given to the intent so then the correct message will be open each time. And `Intent.FLAG_ACTIVITY_CLEAR_TOP` makes sure it created clears the Activity if it already exists at the top of the stack and create it again.

MAILAND-2355