Avoid logout when changing password after key migration happened
When trying to change the mailbox or account password right after a key
migration happened, the API request would fail asking the user to logout
and login again.

This happened because of the `UserKey` param not being set when
performing the API request (eg. `key/private`), while after key
migration happened that param is always expected to be set (and `Keys`
isn't anymore, which will be tackled in the next commit).

Since when key migration happens the app receives an event, by
"replicating" the logic that exists in `LoginService#setAccountMigrationStatus`
also in the eventHandler we can check for the migration and set the
`legacyAccount` flag to false if the migration happened.

This flag will then be used in the ChangePassword job to define whether
params such as `UserKey` should be set to the API request.

Please note that compared to the logic currently implemented in the
`LoginService`, the one introduced in the events have been corrected to
check the keys of all of the user's addresses, since that is the
condition that tells us whether the migration did happen.

MAILAND-2401
MAILAND-2402