Fixes the repeated emissions of create, update and delete label events.
The issue was using the replay param for create/update flow and livedata for
delete events. Bot were replaced with a flow with an extrabuffercapacity param
set to 1, in order to be able to emit values with tryEmit while avoiding the
extra emissions when starting to observe the data.

MAILAND-2423