Fix crash after updating labels/folders
The crash was happening in the event system and the reason was that labels in the arriving events had a null id. This was happening because in the EventResponse we were using a LabelApiModel class to describe a label, and this class is using kotlin serialization, while EventResponse is using Gson. As a solution, a new class was created, LabelEventModel, that uses Gson and can be used in the EventResponse. Additionally, a LabelEventApiMapper was created in order to handle LabelEventModel to LabelApiModel mapping.

MAILAND-2416