Attempt testing usage of address token as a passphrase
This testing attempt failed mainly due to the static usage of
`GoOpenPgp` library in the subject.

Making this an instrumentation test using real keys was also tried, but
resulted in a flacky test that gave no value.

Other alternatives that were evaluated got discarded due to the negative
effort / value tradeoff (eg. extracting OpenPgpCrypto lib behind a local
interface to make it testable).

The test will be pushed as ignored for the purpose of gathering any
feedback / idea that might have been missed so far, will be dropped
afterwards.

MAILAND-2401