Delete `User.isLegacy` and AddressKeyActivationWorker because unused
This class was not used since the integration of the login core modules
few months back.
We confirmed it can be dropped as it won't be used anymore after the key
migration takes place (was used for non-private, sub users only).

The user's property `isLegacy` (called `legacyAccount` on the "old" user
model) was only still used by this worked and it can be dropped too.

MAILAND-2401