Fixes crashes while swiping on messages
First crash was caused by passing a null location id to one of the swipe
actions handlers which are all in kotlin now and all take non nullable type.
The second crash was caused by passing a label repository to the jobs
directly instead of using DI for it- jobs are written to disk and need
to be serializable, which in turn means repository would need to be
serializable.

MAILAND-2439