Do not allow editing drafts locally if their remote body can't be fetched
This is needed to avoid a state for which opening a draft which body
isn't available in the local DB while in offline mode would show an
infinite loading and force the user to choose whether to save the draft
or not when exiting the composer (both choices would cause the draft's
body loss, on one side because the draft is deleted and on the other
because overridden with the local, empty body available in the
composer).

The change was to move the check for "no connectivity" to the
constructor of the `FetchDraftDetailJob` instead of the `onRun` method
(onRun would not be called when offline due to the `requireNetwork`
constraint of this Job).
When the "failed fetching draft body" event is posted on bus, the
composer reacts by showing an error and closing the itself when user ack

MAILAND-2448