Messages sent to self do not show up in the inbox
The issue here was that when the event update for the message comes and the labels now also include INBOX as a location, when we resolve the location we never resolve to INBOX because we always start from the back of the list. So SENT is always the first one that we find. Now simply starting from the top of the list fixes the because also in SENT folder we observe the messages by Label and not by Location.

MAILAND-2546