Fix sending message with wrong attachmentIds
We split the handling of attachment that are already uploaded, we skip the upload in the worker and we don't upload them again. And for the attachment that are not uploaded but have missing needed data like `filePath` in this case we now fail permanantly the worker, and the whole flow after it, chained to it.

MAILAND-2599