Fix the appearing and disappearing of reply buttons in details screen
This commit should fix the following issues:
- reply buttons shown before message body is loaded and clicking them
results in a draft with empty body;
- reply buttons are sometimes not shown at all even though the message
body is loaded;
- the action sheet of a message can be summoned in conversation mode
before the message is loaded and we can click the reply buttons from
there which, once again, results in a draft with empty body.

After this commit the behaviour should be such that reply actions and
the button for summoning the in-message action sheet should be
invisible in the cases where the message is not yet loaded and the
message is a draft.

MAILAND-2323
MAILAND-2540