Fix being forcibly switched to another account sometimes
When opening the app from a notification for multiple new messages in
the current secondary account (acc2) and then switching to another
account (acc1) would result into switching back to the previous account
(acc2) whenever onResume is called on MailboxActivity. This caused
"Failed getting message details" and a decryption error in some cases.
The reason was the removal of one extra (the user id) in checkUserId,
which is called in onResume, was not reflected in the intent.extras.
The solution was to call get and remove directly on the intent, instead
of on the extras.

MAILAND-2597