Fix crash on launching app when offline without accounts logged in
Since MailboxActivity is now the main activity, it is launched before we
log in. Requiring the user when showing the 'no connection' snack bar
was the reason for the crash. The solution is to make it possible to
pass a null user as an argument when creating the no connection snack
bar.

MAILAND-2621