Fix events not arriving every 30 seconds when app is in foreground
The problem arises when we put the app in background, wait for an event
to arrive and then bring the app back in foreground. When the app is in
background the events are fetched every 30 mins. When the app is brought
back in foreground, the AlarmReceiver is not being called in order to
change the frequency with which the events should be fetched. In this
scenario the user would have to wait for 30 mins before any events
arrive, even when the app is in foreground. The fix includes bringing
back the calling of AlarmReceiver from onResume in NavigationActivity,
which had been removed in another commit.

MAILAND-2607