Logs the user out when username not found for user id
If we detect an inconsistent state in the app, in which a username
for the primary id is not present in the shared prefs, log the
user out and log the error. The correct username for that user id
will be put there on the next log in.

Additionally, removes the code setting up alarm and fetching events from
the navigation activity when a new id arrives- it's no longer needed
there and was a leftover from a merge conflict from a previous MR.

MAILAND-2644