Brings back fetching events when new primary user id is received
This is needed to avoid the mailbox being refreshed when the first
scheduled events come in. For some reason the first events received
after logging in refresh the mailbox. If we fetch the updates
immediately after logging in, this is masked by the initial loading
of the mailbox. Requires further investigation in MAILAND-2654.

MAILAND-2644