Fix delete action not working properly when done from the details screen
The issue is that when a message gets deleted, even though that action
returns the user to the mailbox, the view model isn't destroyed on time
and it keeps observing the message in DB. So, when the message is
deleted optimistically from the DB, the observer triggers the getMessage
method which tries to fetch the message again and most of the time
succeeds because the action is not yet registered on the API side. The
fix includes canceling the Job of the Flow which observes and emits the
changes to the message/conversation when the delete action is triggered.

MAILAND-1834